Qt6: bring back translations
Use LinguistTools according to qt-6 localization guide. Since the .qm
files are external and not embedded the translations will likely only
work for UNIX. But it seems that that was the case also before.

Also drop empty translation files that have probably never been used and
have been added accidentially.